requirements.txt
fastapi
uvicorn
gunicorn
streamlit
supabase
postgrest
google-genai
pydantic-settings
python-dotenv
jinja2
